Fix -Wshorten-64-to-32 in self extracting executable

+++ utils/self-extracting-executable/decompressor.cpp
@@ -514,7 +514,7 @@ int main(int/* argc*/, char* argv[])
             return 1;
         }
 
-        if (chmod(self, decompressed_umask))
+        if (chmod(self, static_cast<uint32_t>(decompressed_umask)))
         {
             perror("chmod");
             return 1;
